Make script single js file
