+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "packages": ["app", "packages/*", "scripts/i18n"],
+  "packages": ["app", "packages/*"],
   "version": "independent",
   "npmClient": "yarn",
   "useWorkspaces": true
