+++ package.json
@@ -113,12 +113,18 @@
     "redux-testkit": "^1.0.6",
     "redux-thunk": "^2.4.1",
     "tsd": "^0.19.1",
-    "typescript": "~4.5.4"
+    "typescript": "~4.5.4",
+    "glob": "^8.0.1",
+    "yargs": "^17.4.1",
+    "@types/vinyl-fs": "^2.4.12",
+    "@types/yargs": "^17.0.10",
+    "i18next-parser": "^5.4.0",
+    "vinyl": "^2.2.1",
+    "vinyl-fs": "^3.0.3"
   },
   "workspaces": [
     "app",
-    "packages/*",
-    "scripts/i18n"
+    "packages/*"
   ],
   "packageManager": "yarn@1.18.0"
 }
