--- packages/i18n/locales/core/app/en.json
@@ -1,38 +1,0 @@
-{
-  "Active": "",
-  "Add inventory via CSV": "",
-  "Administration": "",
-  "An error occurred": "",
-  "Card Support": "",
-  "Care Teams Management": "",
-  "Complete": "",
-  "Download Client Data": "",
-  "Draft": "",
-  "Draft Files": "",
-  "Form Configuration": "",
-  "Inventory": "",
-  "JSON Validators": "",
-  "Location Management": "",
-  "Location unit group": "",
-  "Location units": "",
-  "Login": "",
-  "Logout": "",
-  "Manage account": "",
-  "Manifest Releases": "",
-  "Missions": "",
-  "Plans": "",
-  "Please log in with one of the following providers": "",
-  "Product Catalogue": "",
-  "Questionnaire": "",
-  "Retired": "",
-  "Server Settings": "",
-  "Service point inventory": "",
-  "Team Assignment": "",
-  "Team Management": "",
-  "Teams": "",
-  "User Groups": "",
-  "User Management": "",
-  "User Roles": "",
-  "Users": "",
-  "Welcome to OpenSRP": ""
-}
