--- scripts/i18n/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@opensrp/scripts",
-  "version": "0.0.1",
-  "description": "Automation scripts for repo-wide management",
-  "bin": {
-    "opensrp": "bin/cli.js"
-  },
-  "files": [
-    "bin"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": "https://github.com/opensrp/web",
-  "config": {
-    "folderName": "scripts"
-  },
-  "scripts": {
-    "test": " cd ../../ && yarn test $npm_package_config_folderName --verbose --collectCoverage=true && cd $npm_package_config_folderName",
-    "tsc": "tsc",
-    "lint": "eslint ./**/*.{js,jsx,ts,tsx}",
-    "build": "yarn run tsc"
-  },
-  "jest": {
-    "automock": false,
-    "setupFiles": [
-      "../../setupTests"
-    ]
-  },
-  "bugs": {
-    "url": "https://github.com/opensrp/web/issues"
-  },
-  "author": "OpenSRP Engineering",
-  "license": "Apache-2.0",
-  "private": false,
-  "dependencies": {
-    "glob": "^8.0.1",
-    "yargs": "^17.4.1"
-  },
-  "devDependencies": {
-    "@types/vinyl-fs": "^2.4.12",
-    "@types/yargs": "^17.0.10",
-    "i18next-parser": "^5.4.0",
-    "typescript": "^4.7.2",
-    "vinyl": "^2.2.1",
-    "vinyl-fs": "^3.0.3"
-  }
-}
