--- scripts/i18n/src/index.d.ts
@@ -1,2 +1,0 @@
-declare module 'i18next-parser';
-declare module 'yargs/helpers';
