--- scripts/i18n/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "bin",
-    "declaration": true,
-    "declarationDir": "bin/types",
-    "esModuleInterop": true,
-    "emitDeclarationOnly": false
-  },
-  "include": ["src"],
-  "exclude": ["**/node_modules", "**/bin", "**/dist", "**/*.spec.ts", "**/*tests"]
-}
